Refactor utilities to separate files
